+++ Solutions/PowerShell.Commands/Commands/Web/RemoveFile.cs
@@ -1,0 +1,37 @@
+﻿using System.IO;
+using System.Management.Automation;
+using Microsoft.SharePoint.Client;
+using OfficeDevPnP.PowerShell.CmdletHelpAttributes;
+
+namespace OfficeDevPnP.PowerShell.Commands
+{
+    [Cmdlet(VerbsCommon.Remove, "SPOFile")]
+    [CmdletHelp("Removes a file.")]
+    [CmdletExample(Code = @"
+PS:>Remove-SPOFile -ServerRelativeUrl /sites/project/_catalogs/themes/15/company.spcolor", SortOrder = 1)]
+  
+    public class RemoveFile : SPOWebCmdlet
+    {
+        [Parameter(Mandatory = true, Position=0, ValueFromPipeline=true)]
+        public string ServerRelativeUrl = string.Empty;
+
+        [Parameter(Mandatory = false)]
+        public SwitchParameter Force;
+
+        protected override void ExecuteCmdlet()
+        {
+            var file = SelectedWeb.GetFileByServerRelativeUrl(ServerRelativeUrl);
+
+            ClientContext.Load(file, f => f.Name);
+            ClientContext.ExecuteQueryRetry();
+
+            if (Force || ShouldContinue(string.Format(Properties.Resources.Delete0, file.Name), Properties.Resources.Confirm))
+            {
+                file.DeleteObject();
+
+                ClientContext.ExecuteQueryRetry();
+            }
+
+        }
+    }
+}